requirements.txt
beautifulsoup4==4.13.4
docx2txt==0.9
faiss-cpu==1.11.0
fastapi==0.115.12
Jinja2==3.1.6
langchain==0.3.25
langchain-community==0.3.24
openai==1.82.1
pdfminer.six==20250506
pypdf==5.5.0
python-dotenv==1.1.0
sentence-transformers==4.1.0
streamlit==1.45.1
tiktoken==0.9.0
uvicorn==0.34.2